Finish re-factor to functional components and update tests
