+++ packages/venia-concept/src/RootComponents/Category/index.js
@@ -3,4 +3,5 @@
  * description = 'Basic Category Page'
  * pageTypes = CATEGORY
  */
-export { default } from './category-new';
+export { default as Category } from './category';
+export { default } from './container';
